Moved Chat to separate file, Friend notifies listeners on new message
