+++ lib/data.dart
@@ -1,64 +1,9 @@
 import 'dart:io';
-import 'package:mutex/mutex.dart';
 import 'package:flutter/material.dart';
-import 'package:text_messenger/text_widgets.dart';
+import 'package:mutex/mutex.dart';
 
 const int ourPort = 8888;
 final m = Mutex();
-
-class ChatScreen extends StatefulWidget {
-  const ChatScreen({super.key, required this.friend});
-
-  final Friend? friend;
-
-  @override
-  State<ChatScreen> createState() => _ChatScreenState();
-}
-
-class _ChatScreenState extends State<ChatScreen> {
-  late TextEditingController _sendController;
-
-  void initState() {
-    super.initState();
-    _sendController = TextEditingController();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    String currentFriend = widget.friend!.name;
-    return Scaffold(
-      appBar: AppBar(
-        title: Text("Chat with " + widget.friend!.name),
-      ),
-      body: Column(
-        mainAxisAlignment: MainAxisAlignment.center,
-        children: <Widget>[
-          ScrollText(text: widget.friend!.history()),
-          ActionText(
-              width: 200,
-              label: "Send to $currentFriend",
-              inType: TextInputType.text,
-              controller: _sendController,
-              handler: send),
-        ],
-      ),
-    );
-  }
-
-  Future<void> send(String msg) async {
-    await widget.friend!.send(msg).then((value) {
-      setState(() {
-        _sendController.text = "";
-      });
-    })
-        // TODO FIX WITH SNACKBAR
-        .catchError((e) {
-      ScaffoldMessenger.of(context).showSnackBar(SnackBar(
-        content: Text("Error: $e"),
-      ));
-    });
-  }
-}
 
 class Friends extends Iterable<String> {
   Map<String, Friend> _names2Friends = {};
@@ -89,7 +34,7 @@ class Friends extends Iterable<String> {
   Iterator<String> get iterator => _names2Friends.keys.iterator;
 }
 
-class Friend {
+class Friend extends ChangeNotifier {
   final String ipAddr;
   final String name;
   final List<Message> _messages = [];
@@ -108,8 +53,10 @@ class Friend {
   }
 
   Future<void> _add_message(String name, String message) async {
-    await m.protect(
-        () async => _messages.add(Message(author: name, content: message)));
+    await m.protect(() async {
+      _messages.add(Message(author: name, content: message));
+      notifyListeners();
+    });
   }
 
   String history() => _messages
