+++ pubspec.yaml
@@ -26,6 +26,7 @@ dependencies:
 
   mutex: ^3.0.0
   network_info_plus: ^2.1.4+1
+  provider: ^6.0.0
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
